+++ style.css
@@ -38,12 +38,9 @@ body {
 }
 .modal-backdrop {
     opacity:0.85 !important;
-    /* -webkit-backdrop-filter: blur(15px); Safari */
-    /* backdrop-filter: blur(15px); Other browsers */
-    background-color: rgba(255, 255, 255, 0.4);
-    -webkit-backdrop-filter: blur(20px);
-    backdrop-filter: blur(20px);
+    background-color: rgba(0, 0, 0, 0.85);
     z-index: 50;
+    filter:blur(2px);
 }
 .toast-container {
   z-index: 150;
